papiea-python-sdk/requirements.txt
aiohttp==3.6.2
PyYAML==5.3.1